Fix build for rustc changes.
Previously, PreparedStatement did not actually use its 'db lifetime, so
the lifetime checks behaved as if there was no bound on the lifetime. A
rustc change made that a compile error. Adding a PhantomData there, and so
making the lifetime have to be respected, make a number of other changes
necessary. There are now separate lifetime parameters for statements,
result sets, and rows. Not having separate lifetime parameters would cut
off, (for example) a PreparedStatement's lifetime when it got executed,
even after the ResultSet that came from that execute went out of scope. I
believe that was because the lifetimes got somehow fused together when the
ResultSet was made.

A second change is that prepare() takes a reference instead of a mutable
reference. The compiler was treating the PreparedStatement that came from
the prepare as if it kept ownership of the DatabaseConnection without
that. It seems like there ought to be a way around this, but I didn't find
it.